Match ATC Codes to Drug Names

This function matches ATC codes in the input data frame to corresponding drug names using a reference data frame.

The reference data frame must have drug names in the first column and ATC codes in the second column.

Usage

atc2drug(input_data, atc_column, reference_data)

Arguments

input_data

A data frame containing a column of ATC codes to be matched.

atc_column

The name of the column in input_data containing ATC codes.

reference_data

A data frame with drug names in the first column and corresponding ATC codes in the second column.

Value

A modified data frame with an additional column:

atc_code_drug

Drug name matched to each ATC code.

Examples

# Example usage:
data(example_atc_data)
data(example_reference_data)
atc2drug(example_atc_data, "ATC_code", example_reference_data)
#> 
#> Attaching package: ‘dplyr’
#> The following objects are masked from ‘package:stats’:
#> 
#>     filter, lag
#> The following objects are masked from ‘package:base’:
#> 
#>     intersect, setdiff, setequal, union
#>     ATC_code ATC_code_drug
#> 1    B01AC06       aspirin
#> 2    N02BE01   paracetamol
#> 3       <NA>          <NA>
#> 4    M01AE01     ibuprofen
#> 5    R06AE07    cetirizine
#> 6 ABCDE12345          <NA>
#> 7 abcde12345          <NA>
#> 8       <NA>          <NA>