Perform Drug Matching with Stripping Trailing Characters

This function matches drug names in the input data frame to corresponding ATC codes using a reference data frame, with stripping trailing characters such as "mg", "g", etc.

The reference data frame must have drug names in the first column and ATC codes in the second column.

Usage

drug_match_strip(input_data, drug_column, reference_data)

Arguments

input_data

A data frame containing a column of drug names to be matched.

drug_column

The name of the column in input_data containing drug names.

reference_data

A data frame with drug names in the first column and corresponding ATC codes in the second column.

Value

A modified data frame with additional columns:

drug_name_ATC

ATC code matched to each drug name.

drug_name_matched

Closest matching drug name from reference data.

drug_name_flag

Flag indicating mismatch (1) or match (0) between drug name and ATC code prefix.

Examples

# Example usage:
data(example_drug_data_with_mg)
data(example_reference_data)
drug_match_strip(example_drug_data_with_mg, "drug_name", example_reference_data)
#>            drug_name drug_name_stripped drug_name_matched drug_name_flag
#> 1      Aspirin 100mg            aspirin           aspirin              0
#> 2  Paracetamol 500mg        paracetamol       paracetamol              0
#> 3    Ibuprofen 200mg          ibuprofen         ibuprofen              0
#> 4    Cetirizine 10mg         cetirizine        cetirizine              0
#> 5    Lisinopril 20mg         lisinopril        lisinopril              0
#> 6   Metformin 1000mg          metformin         metformin              0
#> 7  Atorvastatin 40mg       atorvastatin      atorvastatin              0
#> 8    Omeprazole 20mg         omeprazole        omeprazole              0
#> 9     Salbutamol 2mg         salbutamol        salbutamol              0
#> 10 Amoxicillin 500mg        amoxicillin       amoxicillin              0
#>    drug_name_ATC
#> 1        B01AC06
#> 2        N02BE01
#> 3        M01AE01
#> 4        R06AE07
#> 5        C09AA03
#> 6        A10BA02
#> 7        C10AA05
#> 8        A02BC01
#> 9        R03AC02
#> 10       J01CA04