Validate ATC Codes Against a Reference List

This function checks whether ATC codes in a given column exist in a reference list of valid ATC codes.

Usage

validate_atc_with_reference(input_data, atc_column, reference_atc_codes)

Arguments

input_data

A data frame containing ATC codes.

atc_column

The name of the column in input_data containing ATC codes.

reference_atc_codes

A character vector of valid ATC codes.

Value

A data frame with an additional column indicating invalid ATC codes:

*_invalid

1 if the ATC code is not in the reference list, 0 if it is.

Examples

data(example_atc_data_for_validation)
data(example_reference_data)

valid_codes <- example_reference_data$ATC_code
validate_atc_with_reference(example_atc_data_for_validation, "atc_1", valid_codes)
#>   ID   atc_1 atc_1_invalid   atc_2
#> 1  1 B01AC06             0 C10AA05
#> 2  2 N02BE01             0 M01AE01
#> 3  3  FAKETC             1 NOPE123
#> 4  4 WRONG01             1 XXXXX01
#> 5  5 Z99ZZ99             1 J01CA04
validate_atc_with_reference(example_atc_data_for_validation, "atc_2", valid_codes)
#>   ID   atc_1   atc_2 atc_2_invalid
#> 1  1 B01AC06 C10AA05             0
#> 2  2 N02BE01 M01AE01             0
#> 3  3  FAKETC NOPE123             1
#> 4  4 WRONG01 XXXXX01             1
#> 5  5 Z99ZZ99 J01CA04             0